separating out NavigableMap features from base features of Trie
